Moved `Flip` and `Texture` to `glium_texture` module
Closes https://github.com/PistonDevelopers/glium_graphics/issues/100
